+++ orte/libjorte/JORTEPublicationWaitForSubscriptions.c
@@ -39,7 +39,7 @@
 
 JNIEXPORT jint JNICALL
 Java_org_ocera_orte_Publication_jORTEPublicationWaitForSubscriptions
-(JNIEnv *env, jobject obj, jint j_appDomain_handle,
+(JNIEnv *env, jobject obj, jlong j_appDomain_handle,
  jobject obj_ntpT, jlong jretries, jlong jnoSubscriptions)
 {
   int            i;
